requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==41.0.7
exceptiongroup==1.2.2
fastapi==0.104.1
filelock==3.17.0
h11==0.14.0
idna==3.10
mangum==0.17.0
packaging==24.2
platformdirs==3.11.0
pycparser==2.22
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.10.1
pyOpenSSL==23.3.0
python-dotenv==1.0.0
pytz==2025.1
requests==2.32.3
sniffio==1.3.1
snowflake-connector-python==3.5.0
sortedcontainers==2.4.0
starlette==0.27.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.24.0